code file split and refactor
