+++ src/actions/types.ts
@@ -1,0 +1,5 @@
+export interface VerificationResult {
+  verified: boolean;
+  actualData: any;
+  message?: string;
+}
